requirements.txt
# Flask and Core Dependencies
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
urllib3==2.0.4

# Authentication and Security
PyJWT==2.8.0
cryptography==41.0.4
bcrypt==4.0.1

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Social Media APIs
tweepy==4.14.0
python-linkedin==2.2.0

# AI and Content Generation (Future)
openai==0.28.1
anthropic==0.3.11

# Background Tasks (Future)
celery==5.3.1
redis==4.6.0

# Email (Future)
Flask-Mail==0.9.1

# File Handling
Pillow==10.0.0

# Utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7

# Development Dependencies
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Production Dependencies
gunicorn==21.2.0
whitenoise==6.5.0

# Monitoring and Logging
sentry-sdk[flask]==1.32.0

# Rate Limiting
Flask-Limiter==3.5.0

# Caching
Flask-Caching==2.1.0

# Webhook Security
hmac==0.0.1

# JSON Web Tokens
PyJWT==2.8.0

# Timezone handling
pytz==2023.3

# URL parsing
urllib3==2.0.4

# Data validation
marshmallow==3.20.1

# Async support (Future)
aiohttp==3.8.5
asyncio==3.4.3

# Testing
factory-boy==3.3.0
faker==19.6.2

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0
